+++ src/main/java/be/baur/sdt/xpath/SDAXPath.java
@@ -9,6 +9,7 @@ import be.baur.sdt.xpath.function.CompareNumberFunction;
 import be.baur.sdt.xpath.function.CompareStringFunction;
 import be.baur.sdt.xpath.function.CurrentDateTimeFunction;
 import be.baur.sdt.xpath.function.DateTimeFunction;
+import be.baur.sdt.xpath.function.DateTimeToLocalFunction;
 import be.baur.sdt.xpath.function.DateTimeToMillisFunction;
 import be.baur.sdt.xpath.function.DateTimeToTimeZoneFunction;
 import be.baur.sdt.xpath.function.DocumentNodeFunction;
@@ -47,8 +48,9 @@ public class SDAXPath extends BaseXPath {
 		XPathFunctionContext ctx = (XPathFunctionContext) XPathFunctionContext.getInstance();
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "compare-number", new CompareNumberFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "compare-string", new CompareStringFunction());
-		ctx.registerFunction(SDT.W3CFUNCTIONS_NS_URI, "current-dateTime", new CurrentDateTimeFunction());
+		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "current-dateTime", new CurrentDateTimeFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "dateTime", new DateTimeFunction());
+		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "dateTime-to-local", new DateTimeToLocalFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "dateTime-to-millis", new DateTimeToMillisFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "dateTime-to-timezone", new DateTimeToTimeZoneFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "document-node", new DocumentNodeFunction());
@@ -59,9 +61,9 @@ public class SDAXPath extends BaseXPath {
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "parse-sda", new ParseSDAFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "render-sda", new RenderSDAFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "right", new RightFunction());
+		ctx.registerFunction(SDT.W3CFUNCTIONS_NS_URI, "string-join", new StringJoinFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "timestamp", new TimestampFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "tokenize", new TokenizeFunction());
-		ctx.registerFunction(SDT.W3CFUNCTIONS_NS_URI, "string-join", new StringJoinFunction());
 	}
 	
 	/**