+++ src/main/java/be/baur/sdt/xpath/function/DateTimeToLocalFunction.java
@@ -1,0 +1,72 @@
+package be.baur.sdt.xpath.function;
+
+import java.time.ZonedDateTime;
+import java.time.temporal.TemporalAccessor;
+import java.util.List;
+
+import org.jaxen.Context;
+import org.jaxen.Function;
+import org.jaxen.FunctionCallException;
+import org.jaxen.Navigator;
+
+/**
+ * <code><i>date-time</i> dateTime-to-local( <i>date-time</i> )</code><br>
+ * <p>
+ * Removes the time zone component from a date-time and returns a local
+ * date-time with the same year, month, day and time as the one supplied.
+ * <p>
+ * Examples:
+ * <p>
+ * <code>sdt:dateTime-to-local('1970-01-01T00:00:00Z')</code> returns
+ * <code>1970-01-01T00:00:00</code>.<br>
+ */
+public class DateTimeToLocalFunction implements Function
+{
+	static String FUNC = "dateTime-to-local()";
+
+    /**
+     * Create a new <code>DateTimeToLocalFunction</code> object.
+     */
+    public DateTimeToLocalFunction() {}
+ 
+
+	/**
+	 * Removes the time zone component from a date-time.
+	 * 
+	 * @param context the context at the point in the expression when the function
+	 *                is called
+	 * @param args    an argument list that contains one item.
+	 * @return a local date-time string
+	 * @throws FunctionCallException if <code>args</code> has more or less than one
+	 *                               item or evaluation failed.
+	 */
+    @Override
+	@SuppressWarnings("rawtypes")
+	public Object call(Context context, List args) throws FunctionCallException {
+
+		if (args.size() != 1)
+			throw new FunctionCallException(FUNC + " requires exactly one argument.");
+
+		return evaluate(args.get(0), context.getNavigator());
+	}
+
+
+	/**
+	 * Removes the time zone component from a date-time.
+	 * 
+	 * @param dtm a local or zoned date-time string
+	 * @param nav the navigator used
+	 * @return a local date-time string
+	 * @throws FunctionCallException if no valid dateTime or time zone was supplied
+	 *                               or conversion to the target time zone failed.
+	 */
+	public static String evaluate(Object dtm, Navigator nav) throws FunctionCallException {
+
+		TemporalAccessor temporal = DateTimeFunction.evaluate(FUNC, dtm, nav);
+
+		if (temporal instanceof ZonedDateTime)
+			temporal = ((ZonedDateTime) temporal).toLocalDateTime();
+
+		return DateTimeFunction.format(temporal);
+	}
+}