+++ src/main/java/be/baur/sdt/xpath/function/DateTimeToMillisFunction.java
@@ -23,7 +23,8 @@ import org.jaxen.Navigator;
  */
 public class DateTimeToMillisFunction implements Function
 {
-
+	static String FUNC = "dateTime-to-millis()";
+	
     /**
      * Create a new <code>DateTimeToMillisFunction</code> object.
      */
@@ -49,7 +50,7 @@ public class DateTimeToMillisFunction implements Function
 		if (args.size() == 1)
 			return evaluate(args.get(0), context.getNavigator());
 
-		throw new FunctionCallException("dateTime-to-millis() requires exactly one argument.");
+		throw new FunctionCallException(FUNC + " requires exactly one argument.");
 	}
 
 
@@ -65,15 +66,15 @@ public class DateTimeToMillisFunction implements Function
 	 */
 	public static Double evaluate(Object obj, Navigator nav) throws FunctionCallException {
 
-		TemporalAccessor dtm = DateTimeFunction.evaluate("dateTime-to-millis()", obj, nav);
+		TemporalAccessor dtm = DateTimeFunction.evaluate(FUNC, obj, nav);
 
 		if (!(dtm instanceof ZonedDateTime))
-			throw new FunctionCallException("dateTime-to-millis() requires a zoned date-time.");
+			throw new FunctionCallException(FUNC + " requires a zoned date-time.");
 
 		try {
 			return (double) Instant.from(dtm).toEpochMilli();
 		} catch (Exception e) {
-			throw new FunctionCallException("dateTime-to-millis() conversion of '" + dtm + "' failed.", e);
+			throw new FunctionCallException(FUNC + " conversion of '" + dtm + "' failed.", e);
 		}
 	}
 }